got rid of some compilation warnings regarding the size_t; fragments now imported upon pasting; some re-factoring;

+++ trunk/repository/IsoSurface.h
@@ -40,7 +40,7 @@ public:
 protected:
 ////////////////////////////////////////////////////////////////////////
   struct IsoPoint {
-    int newID;
+    size_t newID;
     float x, y, z;
     bool initialised;
     IsoPoint(): initialised(false) {  }
